+++ src/libres/db/models/base.py
@@ -1,11 +1,22 @@
 from __future__ import annotations
 
+from datetime import datetime
+from sqlalchemy.orm import registry
+from sqlalchemy.orm import DeclarativeBase
+from uuid import UUID as PythonUUID
 
-from typing import TYPE_CHECKING
-if TYPE_CHECKING:
-    from sqlalchemy.ext.declarative import declarative_base
-else:
-    # FIXME: Move this import out when dropping 1.4 support
-    from sqlalchemy.orm import declarative_base
+from .types import JSON
+from .types import UTCDateTime
+from .types import UUID
 
-ORMBase = declarative_base()
+
+from typing import Any
+
+
+class ORMBase(DeclarativeBase):
+
+    registry = registry(type_annotation_map={
+        datetime: UTCDateTime(timezone=False),
+        dict[str, Any]: JSON,
+        PythonUUID: UUID,
+    })